1.更新项目周报中添加成员计划周报 2.更新时间选择界面 3.更新具体事项，输出物等输入框大小

+++ app/src/main/java/com/aorise/weeklyreport/bean/ProjectPlan.java
@@ -30,7 +30,7 @@ public class ProjectPlan extends BaseObservable {
 
     private int id;
     private String name;
-    private int workTime;
+    private float workTime;
     private int percentComplete;
     private List<OwnerListBean> ownerList;
 
@@ -50,11 +50,11 @@ public class ProjectPlan extends BaseObservable {
         this.name = name;
     }
 
-    public int getWorkTime() {
+    public float getWorkTime() {
         return workTime;
     }
 
-    public void setWorkTime(int workTime) {
+    public void setWorkTime(float workTime) {
         this.workTime = workTime;
     }
